Fix termination on consumer disconnect
When the consumer (aka reader of the out fifo pipe) closes the fifo, the
next write to it will cause a SIGPIPE and the ec process terminates with
exit code 141.
To prevent this, make the process ignore the SIGPIPE signal. This is
fine as that is the only place where ec writes to a fifo, and allows
the error to be handled locally instead of in a global signal handler.
In case a write fails because of a broken pipe, close the fd and re-open
it. This will then again block until a consumer connects, and write to
it when it does.

Fixes https://github.com/voice-engine/ec/issues/13