Optimize frame-size
The speex docs [0] mention that the ec calculation is easier when the
frame-size (in samples) is a power of two, so round up the calculated
value to the next power of two. If it already is one nothing changes,
so e.g. 512 will stay 512 and not be rounded up to 1024. Lucky us,
there is already a function doing exactly that ready-to-use.
This also allows to use rates like 44100, which previously failed with
an error "KissFFT: max radix supported is 17".

[0] https://speex.org/docs/manual/speex-manual/node7.html#317